[tests] Remove spurious error log in p2p_segwit.py
Since 3852e9df8e9239e4bd77645c3ed0f6654e10c365, when wait_until() fails,
an error message is logged to the test framework log. This means that if
wait_until() is called inside a try-except with the expectation that it
will fail, a spurious error message is logged.

wait_until() shouldn't be called with the expectation of failure. Fix
that in p2p_segwit.py.